sim:epiphany: Fix nested interrupt RTI behaviour
When "RTI" is called and there are no pending interrupts, PC should be
set to IRET.

When "RTI" is called and there are pending interrupt(s), PC should be
set to the interrupt vector of the highest priority (least significant
bit in IPEND) interrupt.

This makes "nested_interrupts" from "epiphany-examples" pass.